requirements-dev.txt
-r requirements.txt
black==20.8b1
pytest
pytest-django
pytest-watch
